enabled ITM based logging
Since semi-hosting logging is so slow I enabled ITM based logging which
  is much faster and works very well.
Moved the UART init and examples into a library.
Updated the README with instructions on using itmdump to get ITM logs.
Updated the openocd.gdb setup to enble the tpiu in openocd/gdb and
  disabled semi-hosting in the config for now.
Created led init for a heartbeat LED in the main source file.